+++ codex-rs/core/src/tools/spec.rs
@@ -7,7 +7,6 @@ use crate::features::Feature;
 use crate::features::Features;
 use crate::mcp_connection_manager::ToolInfo;
 use crate::models_manager::collaboration_mode_presets::CollaborationModesConfig;
-use crate::tools::handlers::PINNED_ARTIFACT_RUNTIME_VERSION;
 use crate::tools::handlers::PLAN_TOOL;
 use crate::tools::handlers::SEARCH_TOOL_BM25_DEFAULT_LIMIT;
 use crate::tools::handlers::SEARCH_TOOL_BM25_TOOL_NAME;
@@ -35,7 +34,6 @@ use serde_json::Value as JsonValue;
 use serde_json::json;
 use std::collections::BTreeMap;
 use std::collections::HashMap;
-use std::path::Path;
 
 const SEARCH_TOOL_BM25_DESCRIPTION_TEMPLATE: &str =
     include_str!("../../templates/search_tool/tool_description.md");
@@ -77,7 +75,6 @@ pub(crate) struct ToolsConfig {
 }
 
 pub(crate) struct ToolsConfigParams<'a> {
-    pub(crate) codex_home: &'a Path,
     pub(crate) model_info: &'a ModelInfo,
     pub(crate) features: &'a Features,
     pub(crate) web_search_mode: Option<WebSearchMode>,
@@ -87,7 +84,6 @@ pub(crate) struct ToolsConfigParams<'a> {
 impl ToolsConfig {
     pub fn new(params: &ToolsConfigParams) -> Self {
         let ToolsConfigParams {
-            codex_home,
             model_info,
             features,
             web_search_mode,
@@ -102,11 +98,8 @@ impl ToolsConfig {
         let include_default_mode_request_user_input =
             include_request_user_input && features.enabled(Feature::DefaultModeRequestUserInput);
         let include_search_tool = features.enabled(Feature::Apps);
-        let include_artifact_tools = features.enabled(Feature::Artifact)
-            && codex_artifacts::is_js_runtime_available(
-                codex_home,
-                PINNED_ARTIFACT_RUNTIME_VERSION,
-            );
+        let include_artifact_tools =
+            features.enabled(Feature::Artifact) && codex_artifacts::can_manage_artifact_runtime();
         let include_image_gen_tool =
             features.enabled(Feature::ImageGeneration) && supports_image_generation(model_info);
         let include_agent_jobs = include_collab_tools && features.enabled(Feature::Sqlite);
@@ -2174,7 +2167,6 @@ mod tests {
         let mut features = Features::with_defaults();
         features.enable(Feature::UnifiedExec);
         let config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: Path::new("."),
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Live),
@@ -2238,7 +2230,6 @@ mod tests {
         let mut features = Features::with_defaults();
         features.enable(Feature::Collab);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2261,51 +2252,12 @@ mod tests {
     fn test_build_specs_artifact_tool_enabled() {
         let mut config = test_config();
         let runtime_root = tempfile::TempDir::new().expect("create temp codex home");
-        let platform = codex_artifacts::ArtifactRuntimePlatform::detect_current()
-            .expect("detect artifact platform");
-        let install_dir = runtime_root
-            .path()
-            .join("packages")
-            .join("artifacts")
-            .join(PINNED_ARTIFACT_RUNTIME_VERSION)
-            .join(platform.as_str());
-        std::fs::create_dir_all(install_dir.join("node/bin")).expect("create runtime dir");
-        std::fs::create_dir_all(install_dir.join("artifact-tool/dist"))
-            .expect("create build entrypoint dir");
-        std::fs::create_dir_all(install_dir.join("granola-render/dist"))
-            .expect("create render entrypoint dir");
-        std::fs::write(
-            install_dir.join("manifest.json"),
-            serde_json::json!({
-                "schema_version": 1,
-                "runtime_version": PINNED_ARTIFACT_RUNTIME_VERSION,
-                "node": { "relative_path": "node/bin/node" },
-                "entrypoints": {
-                    "build_js": { "relative_path": "artifact-tool/dist/artifact_tool.mjs" },
-                    "render_cli": { "relative_path": "granola-render/dist/render_cli.mjs" }
-                }
-            })
-            .to_string(),
-        )
-        .expect("write manifest");
-        std::fs::write(install_dir.join("node/bin/node"), "#!/bin/sh\n").expect("write node");
-        std::fs::write(
-            install_dir.join("artifact-tool/dist/artifact_tool.mjs"),
-            "export const ok = true;\n",
-        )
-        .expect("write build entrypoint");
-        std::fs::write(
-            install_dir.join("granola-render/dist/render_cli.mjs"),
-            "export const ok = true;\n",
-        )
-        .expect("write render entrypoint");
         config.codex_home = runtime_root.path().to_path_buf();
         let model_info =
             ModelsManager::construct_model_info_offline_for_tests("gpt-5-codex", &config);
         let mut features = Features::with_defaults();
         features.enable(Feature::Artifact);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2324,7 +2276,6 @@ mod tests {
         features.enable(Feature::Collab);
         features.enable(Feature::Sqlite);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2355,7 +2306,6 @@ mod tests {
             ModelsManager::construct_model_info_offline_for_tests("gpt-5-codex", &config);
         let mut features = Features::with_defaults();
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2370,7 +2320,6 @@ mod tests {
 
         features.enable(Feature::DefaultModeRequestUserInput);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2394,7 +2343,6 @@ mod tests {
         let mut features = Features::with_defaults();
         features.disable(Feature::MemoryTool);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2415,7 +2363,6 @@ mod tests {
         let features = Features::with_defaults();
 
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2442,7 +2389,6 @@ mod tests {
         features.enable(Feature::JsRepl);
 
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2465,7 +2411,6 @@ mod tests {
         image_generation_features.enable(Feature::ImageGeneration);
 
         let default_tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &supported_model_info,
             features: &default_features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2480,7 +2425,6 @@ mod tests {
         );
 
         let supported_tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &supported_model_info,
             features: &image_generation_features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2490,7 +2434,6 @@ mod tests {
         assert_contains_tool_names(&supported_tools, &["image_generation"]);
 
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &unsupported_model_info,
             features: &image_generation_features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2526,10 +2469,9 @@ mod tests {
         web_search_mode: Option<WebSearchMode>,
         expected_tools: &[&str],
     ) {
-        let config = test_config();
+        let _config = test_config();
         let model_info = model_info_from_models_json(model_slug);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features,
             web_search_mode,
@@ -2564,7 +2506,6 @@ mod tests {
         let features = Features::with_defaults();
 
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2590,7 +2531,6 @@ mod tests {
         let features = Features::with_defaults();
 
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Live),
@@ -2617,7 +2557,6 @@ mod tests {
         let features = Features::with_defaults();
 
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Live),
@@ -2647,7 +2586,6 @@ mod tests {
             ModelsManager::construct_model_info_offline_for_tests("gpt-5-codex", &config);
         let features = Features::with_defaults();
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2671,7 +2609,6 @@ mod tests {
             ModelsManager::construct_model_info_offline_for_tests("gpt-5-codex", &config);
         let features = Features::with_defaults();
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2863,7 +2800,6 @@ mod tests {
         let mut features = Features::with_defaults();
         features.enable(Feature::UnifiedExec);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Live),
@@ -2888,7 +2824,6 @@ mod tests {
         features.enable(Feature::ShellZshFork);
 
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Live),
@@ -2915,7 +2850,6 @@ mod tests {
         let mut features = Features::with_defaults();
         features.enable(Feature::UnifiedExec);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2932,7 +2866,7 @@ mod tests {
 
     #[test]
     fn test_test_model_info_includes_sync_tool() {
-        let config = test_config();
+        let _config = test_config();
         let mut model_info = model_info_from_models_json("gpt-5-codex");
         model_info.experimental_supported_tools = vec![
             "test_sync_tool".to_string(),
@@ -2942,7 +2876,6 @@ mod tests {
         ];
         let features = Features::with_defaults();
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -2975,7 +2908,6 @@ mod tests {
         let mut features = Features::with_defaults();
         features.enable(Feature::UnifiedExec);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Live),
@@ -3063,7 +2995,6 @@ mod tests {
         let mut features = Features::with_defaults();
         features.enable(Feature::UnifiedExec);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -3110,7 +3041,6 @@ mod tests {
         let mut features = Features::with_defaults();
         features.enable(Feature::Apps);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -3179,7 +3109,6 @@ mod tests {
         let mut features = Features::with_defaults();
         features.enable(Feature::UnifiedExec);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -3235,7 +3164,6 @@ mod tests {
         let mut features = Features::with_defaults();
         features.enable(Feature::UnifiedExec);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -3288,7 +3216,6 @@ mod tests {
         features.enable(Feature::UnifiedExec);
         features.enable(Feature::ApplyPatchFreeform);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -3343,7 +3270,6 @@ mod tests {
         let mut features = Features::with_defaults();
         features.enable(Feature::UnifiedExec);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
@@ -3477,7 +3403,6 @@ Examples of valid command strings:
         let mut features = Features::with_defaults();
         features.enable(Feature::UnifiedExec);
         let tools_config = ToolsConfig::new(&ToolsConfigParams {
-            codex_home: &config.codex_home,
             model_info: &model_info,
             features: &features,
             web_search_mode: Some(WebSearchMode::Cached),
